AzureFunctions/requirements.txt
azure-functions
requests
psycopg2-binary
PyJWT
openpyxl
msal